MedicalCopilot/scripts/pyproject.toml
[tool.poetry]
name = "amina"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
requests = "^2.32.3"


[tool.poetry.group.scripts.dependencies]
deepgram-sdk = "^3.4.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
pyaudio = "^0.2.14"


[tool.poetry.group.script.dependencies]
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
